:sparkles: feat(consensus ): insert data from the consensus control list. LFUB-108

+++ onecgiar-pr-server/src/api/ipsr/result-innovation-package/entities/consensus-initiative-work-package.entity.ts
@@ -5,7 +5,7 @@ import { BaseEntity } from '../../../../shared/entities/base-entity';
 export class consensusInitiativeWorkPackage{
     @PrimaryGeneratedColumn({
         name: 'consensus_initiative_work_package_id',
-        type: 'int'
+        type: 'bigint'
     })
     consensus_initiative_work_package_id: number;
